Refactor imports: consolidate model and service imports, and update widget exports for better organization
