+++ lib/controllers/bindings.dart
@@ -2,11 +2,9 @@ import 'dart:developer';
 
 import 'package:get/get.dart';
 import 'package:http/http.dart' as http;
-import 'package:todo_test_task/controllers/find_controller.dart';
-import 'package:todo_test_task/services/api_service.dart';
-import 'package:todo_test_task/services/storage_service.dart';
-import '../services/todo_service.dart';
-import '../controllers/todo_controller.dart';
+
+import 'package:todo_test_task/controllers/controllers.dart';
+import 'package:todo_test_task/services/services.dart';
 
 class HomeBinding extends Bindings {
   @override
