+++ lib/helpers/external_helpers.dart
@@ -1,19 +1,17 @@
 import 'dart:async';
 import 'package:audioplayers/audioplayers.dart';
 
-enum SoundType { archive, notify, completed }
+import 'package:todo_test_task/common/common.dart';
 
 class ExternalHelpers {
   static final AudioPlayer _player = AudioPlayer();
 
-  /// Map enum to asset path
   static const Map<SoundType, String> _soundAssets = {
     SoundType.archive: 'archived.mp3',
     SoundType.notify: 'notification_sound.mp3',
     SoundType.completed: 'completed.mp3',
   };
 
-  /// Play sound by type
   static Future<void> playSound(SoundType type) async {
     final asset = _soundAssets[type];
     if (asset != null) {
