+++ lib/views/account/account_page.dart
@@ -1,14 +1,10 @@
-// ignore_for_file: unused_local_variable
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+
 import 'package:todo_test_task/common/common.dart';
-import 'package:todo_test_task/controllers/account_controller.dart';
-import 'package:todo_test_task/controllers/todo_controller.dart';
-import 'package:todo_test_task/helpers/ui_helpers.dart';
-import 'package:todo_test_task/models/account.dart';
+import 'package:todo_test_task/controllers/controllers.dart';
+import 'package:todo_test_task/models/models.dart';
 import 'package:todo_test_task/routes.dart';
-import 'package:todo_test_task/widgets/app_network_image.dart';
 import 'package:todo_test_task/widgets/widgets.dart';
 
 class AccountPage extends StatelessWidget {
