+++ lib/views/settings/setting_page.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+
+import 'package:todo_test_task/common/common.dart';
 import 'package:todo_test_task/controllers/ui_controller.dart';
 import 'package:todo_test_task/helpers/ui_helpers.dart';
