+++ lib/widgets/app_network_image.dart
@@ -1,5 +1,5 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 
 class CachedImage extends StatelessWidget {
   final String? url;
