+++ lib/widgets/widgets.dart
@@ -1,4 +1,6 @@
 export 'app_button.dart';
 export 'app_chip.dart';
+export 'app_network_image.dart';
 export 'error_page.dart';
+export 'pro_feature_wrapper.dart';
 export 'todo_detail_sheet.dart';
